requirements.txt
threatexchange>=1.0.0
albumentations>=1.3.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
typer>=0.9.0
PyYAML>=6.0
rich>=13.0.0
